backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis & Caching
redis==5.0.1
aioredis==2.0.1
hiredis==2.3.2

# Background Tasks
celery==5.3.6
flower==2.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2

# AI/ML Services
anthropic==0.18.0
openai==1.10.0
google-generativeai==0.3.2
httpx==0.26.0

# Transcription Services
deepgram-sdk==3.2.0
assemblyai==0.23.0

# Cloud Storage
boto3==1.34.34
google-cloud-storage==2.14.0

# CRM Integrations
simple-salesforce==1.12.5
hubspot-api-client==9.0.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0
pytz==2024.1
python-dateutil==2.8.2
email-validator==2.1.0.post1

# Monitoring & Logging
sentry-sdk==1.40.0
structlog==24.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.6.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
